+++ config/me_cms.php
@@ -1,43 +1,29 @@
 <?php
 return ['MeCms' => [
-	//Backend
-	'backend' => [
-		//Layout you want to use for the backend
-		//If you want to use a layout from your application (eg. `default.ctp`), change this value without extension
-		'layout' => 'MeCms.backend',
+	//Admin layout
+	'admin' => [
 		//Number of photos to show per page. This must be a multiple of 4
 		'photos' => 12,
 		//Number of records to show per page
 		'records' => 10,
 	 ],
-	//Email
-	'email' => [
-		//EmailTransport configuration
-		'config' => 'default',
-		//Address used as the sender for emails sent to users and as a 
-		//recipient for the email sent by users
-		'webmaster' => 'email@example.com',
-	],
-	//Frontend
-	'frontend' => [
-		//Google Analytics ID or FALSE
+	//Default layout
+	'default' => [
+		//Google Analytics ID or `FALSE` to disable
 		'analytics' => FALSE,
-		//It shows the alert for the cookie policy
+		//Displays the alert for the cookie policy
 		'cookies_policy' => TRUE,
-		//Layout you want to use for the frontend
-		//If you want to use a layout from your application (eg. `default.ctp`), change this value without extension
-		'layout' => 'MeCms.frontend',
-		//Contact form (enabled or disabled).
+		//Contact form (enabled or disabled)
 		'contact_form' => TRUE,
-		//Facebook app ID or FALSE
+		//Facebook app ID or `FALSE`
 		'facebook_app_id' => FALSE,
-		//Fancybox for photos
+		//Uses Fancybox for photos
 		'fancybox' => TRUE,
 		//Site logo. Relative path to `APP/webroot/img/`
 		'logo' => 'logo.png',
 		//Site offline (enabled or disabled)
 		'offline' => FALSE,
-		//Text to display when the site is offline
+		//Text to display when the site is offline or `FALSE`
 		'offline_text' => FALSE,
 		//Number of records to show per page
 		'records' => 10,
@@ -45,16 +31,26 @@ return ['MeCms' => [
 		'records_for_rss' => 20,
 		//Number of records to show for searches
 		'records_for_searches' => 20,
-		//Automatically adds the meta tag for RSS resources
+		//Adds automatically the meta tag for RSS resources
 		'rss_meta' => TRUE,
-		//Theme. Must be located in `APP/plugins/`
+		//Theme. Must be located into `APP/plugins/`
 		'theme' => FALSE,
-		//For some mobile browsers you can choose a color for the browser bar. Must be a valid HEX color or FALSE to disable
+		//For some mobile browsers you can choose a color for the browser bar.
+        //Must be a valid HEX color or `FALSE to disable`.
 		//See https://developers.google.com/web/updates/2014/11/Support-for-theme-color-in-Chrome-39-for-Android
 		'toolbar_color' => FALSE,
-		//Number of characters to truncate a text. `0` or `FALSE` to disable
-		//Note that you can use the "<!-- read-more -->" tag to indicate manually where to truncate a text
+		//Number of characters to truncate a text. `0` or `FALSE` to disable.
+		//Note that you can use the "<!-- read-more -->" tag to indicate 
+        //  manually where to truncate a text
 		'truncate_to' => 1000,
+	],
+	//Email
+	'email' => [
+		//EmailTransport configuration
+		'config' => 'default',
+		//Address used as the sender for emails sent to users and as a 
+		//  recipient for the email sent by users
+		'webmaster' => 'email@example.com',
 	],
 	//KCFinder
 	'kcfinder' => [
@@ -63,16 +59,16 @@ return ['MeCms' => [
 	],
 	//Main
 	'main' => [
-		//Date formats
-		//See; http://php.net/manual/it/datetime.formats.php
+		//Date formats.
+		//See: http://php.net/manual/it/datetime.formats.php
 		'date' => [
 			//Long format
 			'long'	=> 'YYYY/MM/dd',
 			//Short format
 			'short'	=> 'yy/MM/dd',
 		],
-		//Datetime formats
-		//See; http://php.net/manual/it/datetime.formats.php
+		//Datetime formats.
+		//See: http://php.net/manual/it/datetime.formats.php
 		'datetime' => [
 			//Long format
 			'long'	=> 'YYYY/MM/dd, HH:mm',
@@ -81,13 +77,10 @@ return ['MeCms' => [
 		],
 		//Forces debug on localhost (enabled or disabled)
 		'debug_on_localhost' => TRUE,
-		//Interface language.
-		//With "auto" value, it will try to use the browser language
-		'language' => 'auto',
         //Sitemap expiration. Must be a valid strtotime string
         'sitemap_expiration' => '+24 hours',
 		//Time formats
-		//See; http://php.net/manual/it/datetime.formats.php
+		//See: http://php.net/manual/it/datetime.formats.php
 		'time' => [
 			//Long format
 			'long'	=> 'HH:mm',
@@ -103,8 +96,9 @@ return ['MeCms' => [
 		'category' => TRUE,
 		//Displays the page created datetime
 		'created' => FALSE,
-		//Displays the Shareaholic social buttons
-		//Remember you have to set app and site IDs. See `shareaholic.app_id` and `shareaholic.site_id`
+		//Displays the Shareaholic social buttons.
+		//Remember you have to set app and site IDs. See `shareaholic.app_id` 
+        //  and `shareaholic.site_id`
 		'shareaholic' => FALSE,
 	],
 	//Posts
@@ -117,35 +111,36 @@ return ['MeCms' => [
 		'created' => TRUE,
 		//Adds post tags as keywords meta-tag
 		'keywords' => TRUE,
-		//Related posts
+		//Related posts. `FALSE` to disable
 		'related' => [
-			//Limit of related posts to get for each post. Use `0` to disable.
+			//Limit of related posts to get for each post.
 			//If you use images, it recommended a multiple of 4 
 			'limit' => 4,
 			//Gets only related posts with images
 			'images' => TRUE,
 		],
 		//Displays the Shareaholic social buttons
-		//Remember you have to set app and site IDs. See `shareaholic.app_id` and `shareaholic.site_id`
+		//Remember you have to set app and site IDs. See `shareaholic.app_id` 
+        //  and `shareaholic.site_id`
 		'shareaholic' => FALSE,
 		//Displays the post tags
 		'tags' => TRUE,
 	],
 	//Security
 	'security' => [
-		//Link for "IP map". The `{IP}` string will be replaced with the IP address
+		//Link for "IP map". The `{IP}` string will be replaced
 		'ip_map' => 'http://www.traceip.net/?query={IP}',
-		//Link for "IP who is". The `{IP}` string will be replaced with the IP address
+		//Link for "IP who is". The `{IP}` string will be replaced
 		'ip_whois' => 'http://www.traceip.net/whois/{IP}',
 		//reCAPTCHA (enabled or disabled).
-		//It will be used for some actions, such as signup or reset the password
+		//It will be used for some actions, such as signup or password reset
 		'recaptcha' => FALSE,
-		//Interval between searches, in seconds. Set to `0` or `FALSE` to disable
+		//Interval between searches, in seconds. `0` or `FALSE` to disable
 		'search_interval' => 10,
 	],
 	//Shareaholic
 	'shareaholic' => [
-		//App ID. Used for render the "share buttons" of Shareaholic
+		//App ID. Used for render the "share buttons" of Shareaholic.
 		//You can found it on the "Sharing: Edit Share Button Location"
 		'app_id' => '',
 		//Site ID. Used for render the "setup code" of Shareaholic.
@@ -154,10 +149,10 @@ return ['MeCms' => [
 	],
 	//Users
 	'users' => [
-		//How to activating accounts
-		//	`0` - No activation required, the account is immediately active
-		//	`1` - The account will be enabled by the user via email (default)
-		//	`2`	- The account will be enabled by an administrator
+		//How to activating accounts:
+		//	`0` - No activation required, the account is immediately active;
+		//	`1` - The account will be enabled by the user via email (default);
+		//	`2`	- The account will be enabled by an administrator.
 		'activation' => 1,
 		//Login with cookies ("remember me" function)
 		'cookies_login' => TRUE,
@@ -165,9 +160,9 @@ return ['MeCms' => [
 		'default_group' => 3,
 		//Reset password (enabled or disabled)
 		'reset_password' => TRUE,
-        //Displays the userbar
-        'userbar' => TRUE,
 		//Signup (enabled or disabled)
 		'signup' => TRUE,
-	]
+        //Displays the userbar
+        'userbar' => TRUE,
+	],
 ]];