+++ src/Controller/Admin/PhotosController.php
@@ -98,7 +98,7 @@ class PhotosController extends AppController {
 		//Sets the paginate limit and the maximum paginate limit
 		//See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
         if($render === 'grid') {
-            $this->paginate['limit'] = $this->paginate['maxLimit'] = config('backend.photos');
+            $this->paginate['limit'] = $this->paginate['maxLimit'] = config('admin.photos');
         }
 		
 		$this->set('photos', $this->paginate($this->Photos->queryFromFilter($query, $this->request->query)));