+++ src/Controller/AppController.php
@@ -31,22 +31,25 @@ use Cake\Network\Exception\InternalErrorException;
 class AppController extends BaseController {
 	/**
      * Checks if the latest search has been executed out of the minimum interval
+     * @param string $query_id Query
 	 * @return bool
 	 */
-	protected function _checkLastSearch($query_id = NULL) {
+	protected function _checkLastSearch($query_id = FALSE) {
         $interval = config('security.search_interval');
 		
-        if(empty($interval)) {
+        if(!$interval) {
             return TRUE;
         }
-		
-		$query_id = empty($query_id) ? NULL : md5($query_id);
+        
+        if($query_id) {
+            $query_id = md5($query_id);
+        }
 		
 		$last_search = $this->request->session()->read('last_search');
 		
-		if(!empty($last_search)) {
+		if($last_search) {
 			//Checks if it's the same search
-			if(!empty($query_id) && !empty($last_search['id']) && $query_id === $last_search['id']) {
+			if($query_id && !empty($last_search['id']) && $query_id === $last_search['id']) {
 				return TRUE;
             }
 			//Checks if the interval has not yet expired
@@ -55,7 +58,10 @@ class AppController extends BaseController {
             }
 		}
 		
-		$this->request->session()->write('last_search', ['id' => $query_id, 'time' => time()]);
+		$this->request->session()->write('last_search', [
+            'id' => $query_id,
+            'time' => time(),
+        ]);
 		
 		return TRUE;
 	}
@@ -103,7 +109,14 @@ class AppController extends BaseController {
 		
 		//Sets the paginate limit and the maximum paginate limit
 		//See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
-		$this->paginate['limit'] = $this->paginate['maxLimit'] = $this->request->isAdmin() ? config('backend.records') : config('frontend.records');
+        if($this->request->isAdmin()) {
+            $this->paginate['limit'] = config('admin.records');
+        }
+        else {
+            $this->paginate['limit'] = config('default.records');
+        }
+        
+		$this->paginate['maxLimit'] = $this->paginate['limit'];
 		
 		parent::beforeFilter($event);
 	}
@@ -118,13 +131,18 @@ class AppController extends BaseController {
 	 * @uses App\Controller\AppController::beforeRender()
 	 */
 	public function beforeRender(\Cake\Event\Event $event) {
-		//Ajax layout
+		//Layout for ajax requests
 		if($this->request->is('ajax')) {
 			$this->viewBuilder()->layout('MeCms.ajax');
         }
 		
 		//Uses a custom View class (`MeCms.AppView` or `MeCms.AdminView`)
-        $this->viewBuilder()->className($this->request->isAdmin() ? 'MeCms.View/Admin' : 'MeCms.View/App');
+        if($this->request->isAdmin()) {
+            $this->viewBuilder()->className('MeCms.View/Admin');
+        }
+        else {
+            $this->viewBuilder()->className('MeCms.View/App');
+        }
         
         //Loads the `Auth` helper.
         //The `helper is loaded here (instead of the view) to pass user data
@@ -171,7 +189,7 @@ class AppController extends BaseController {
 	 * @uses MeTools\Network\Request::isAction()
 	 */
 	protected function isOffline() {
-		if(!config('frontend.offline')) {
+		if(!config('default.offline')) {
 			return FALSE;
         }